Work with date and time moved to every file
