+++ bot/schedule/DisrtibuteSchedule.py
@@ -1,10 +1,18 @@
+import pendulum
 from pendulum import today
 
 from bot.Api import Vk
 from bot.database.DataBases import UserBase
 from bot.schedule.GetSchedule import GetSchedule
 from bot.stuff import Utilities
-from bot.stuff.Utilities import GetScheduleTomorrow
+from bot.stuff.Config import Config
+from bot.stuff.Utilities import FORMAT, TZ
+
+
+def GetScheduleTomorrow(schedule_date=pendulum.tomorrow(TZ)):
+    if Config.REDIRECT_DATE:
+        return Config.REDIRECT_DATE
+    return schedule_date.__format__(FORMAT) if schedule_date.weekday() != 6 else schedule_date.add(days=2).__format__(FORMAT)
 
 
 def ClassSend(cls, date):
